# Remove Path key/subscript to save GPU memory usage to improve perf