Remove Path key/subscript to save GPU memory usage to improve perf.

I found less GPU stack memory usage will help on perf.
I decreased the max path depth from 16 to 8, and get perf improvement.
```
constexpr int max_path_depth = 8;
```
And I optimized the main flow logic to reduced query paths:  
- key and named to one path => named
- subscript and index/wildcard to one path => index/wildcard
From the definition of JSON query path,  key and named always ocurr in the same time, and also subscript and index/wildcard are always occur in the same time. So it's safe to reduce (key, named) to (named),  and it's safe to reduce (subscript, index/wildcard) to (index/wildcard)

Note: If query path is a single `wildcard` , e.g.:  $.* , Spark returns null, so the path is equivalent to invalid path.

## performance test
| Customer   | Base line(1st, 2nd, 3rd) | After this PR  | speed |
|------------|-----------|---------|------------|
| customer 1 | 231500 ms, 232760 ms, 232647 ms | 275719 ms, 272898 ms, 274407 ms |   about 84%  |