@res-life 2024-04-22
Some IT cases are not passed:

```
========= 9 failed, 58 passed, 27086 deselected, 9 warnings in 18.72s ==========
```