@res-life 2024-04-23
All IT cases passed after applying two PRs:

- https://github.com/NVIDIA/spark-rapids/pull/10734
- This PR